Moved state stuff into wasm
