+++ include/infra/infra.h
@@ -6,8 +6,6 @@
 #include <spdlog/spdlog.h>
 
 #include <thread>
-#include <mutex>
-#include <shared_mutex>
 
 #include <proto/faasm.pb.h>
 #include <hiredis/hiredis.h>
@@ -130,82 +128,4 @@ namespace infra {
 
         static std::string getFunctionCounterName(const message::Message &msg);
     };
-
-    typedef std::pair<long, long> Segment;
-    typedef std::set<std::pair<long, long>> SegmentSet;
-
-    class StateKeyValue {
-    public:
-        StateKeyValue(const std::string &keyIn);
-
-        const std::string key;
-
-        std::vector<uint8_t> get();
-
-        std::vector<uint8_t> getSegment(long offset, long length);
-
-        void set(const std::vector<uint8_t> &data);
-
-        void setSegment(long offset, const std::vector<uint8_t> &data);
-
-        void pull();
-
-        void pushFull();
-
-        static SegmentSet mergeSegments(SegmentSet);
-
-        void pushPartial();
-
-        void clear();
-
-        long getLocalValueSize();
-
-    private:
-        util::Clock &clock;
-
-        std::atomic<bool> isWholeValueDirty;
-        std::set<std::pair<long, long>> dirtySegments;
-
-        std::vector<uint8_t> value;
-        std::shared_mutex valueMutex;
-
-        util::TimePoint lastPull;
-        util::TimePoint lastInteraction;
-        long staleThreshold;
-        long idleThreshold;
-
-        std::atomic<bool> isNew;
-
-        void doRemoteRead();
-
-        void updateLastInteraction();
-
-        bool isStale(const util::TimePoint &now);
-
-        bool isIdle(const util::TimePoint &now);
-    };
-
-    typedef std::map<std::string, StateKeyValue *> KVMap;
-    typedef std::pair<std::string, StateKeyValue *> KVPair;
-
-    class State {
-    public:
-        State();
-
-        ~State();
-
-        StateKeyValue *getKV(const std::string &key);
-
-        void pushAll();
-
-        void pushLoop();
-
-    private:
-        KVMap local;
-        std::shared_mutex localMutex;
-
-        long pushInterval;
-    };
-
-    State &getGlobalState();
 };
